move PraosNonce ad GPLeader
