+++ chain-impl-mockchain/src/chaineval.rs
@@ -2,12 +2,44 @@ use crate::certificate::PoolId;
 use crate::chaintypes::ChainLength;
 use crate::date::BlockDate;
 use crate::fragment::BlockContentHash;
-use crate::leadership;
+
+use crate::key::Hash;
+
+/// PraosNonce gathered per block
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct PraosNonce([u8; 32]);
+
+impl PraosNonce {
+    pub fn zero() -> Self {
+        PraosNonce([0u8; 32])
+    }
+
+    pub fn from_output_array(array: [u8; 32]) -> Self {
+        PraosNonce(array)
+    }
+
+    /// Change the nonce to be the result of the hash of the current nonce
+    /// and the new supplied nonce.
+    ///
+    /// Effectively: Self = H(Self, Supplied-Hash)
+    pub fn hash_with(&mut self, other: &Self) {
+        let mut buf = [0; 64];
+        buf[0..32].copy_from_slice(&self.0);
+        buf[32..64].copy_from_slice(&other.0);
+        self.0.copy_from_slice(Hash::hash_bytes(&buf).as_ref())
+    }
+}
+
+impl AsRef<[u8]> for PraosNonce {
+    fn as_ref(&self) -> &[u8] {
+        &self.0[..]
+    }
+}
 
 /// Genesis Praos related data extract from the header
 #[derive(Debug, Clone)]
 pub(crate) struct HeaderGPContentEvalContext {
-    pub(crate) nonce: leadership::genesis::Nonce,
+    pub(crate) nonce: PraosNonce,
     pub(crate) pool_creator: PoolId,
 }
 
@@ -24,6 +56,16 @@ pub struct HeaderContentEvalContext {
 mod test {
     use super::*;
     use quickcheck::{Arbitrary, Gen};
+    use std::iter;
+
+    impl Arbitrary for PraosNonce {
+        fn arbitrary<G: Gen>(g: &mut G) -> Self {
+            let mut nonce = [0; 32];
+            let nonce_vec: Vec<u8> = iter::from_fn(|| Some(u8::arbitrary(g))).take(32).collect();
+            nonce.copy_from_slice(&nonce_vec);
+            PraosNonce(nonce)
+        }
+    }
 
     impl Arbitrary for HeaderGPContentEvalContext {
         fn arbitrary<G: Gen>(g: &mut G) -> Self {
