+++ chain-impl-mockchain/src/key.rs
@@ -5,9 +5,11 @@ use chain_core::mempack::{read_mut_slice, ReadBuf, ReadError, Readable};
 use chain_core::property;
 use chain_crypto as crypto;
 use chain_crypto::{
-    AsymmetricKey, AsymmetricPublicKey, SecretKey, SigningAlgorithm, VerificationAlgorithm,
+    digest::DigestOf, AsymmetricKey, AsymmetricPublicKey, Blake2b256, Curve25519_2HashDH,
+    PublicKey, SecretKey, SigningAlgorithm, SumEd25519_12, VerificationAlgorithm,
 };
 use rand_core::{CryptoRng, RngCore};
+use typed_bytes::ByteBuilder;
 
 use std::str::FromStr;
 
@@ -308,9 +310,41 @@ impl FromStr for Hash {
     }
 }
 
+/// Praos Leader consisting of the KES public key and VRF public key
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct GenesisPraosLeader {
+    pub kes_public_key: PublicKey<SumEd25519_12>,
+    pub vrf_public_key: PublicKey<Curve25519_2HashDH>,
+}
+
+impl GenesisPraosLeader {
+    pub fn digest(&self) -> DigestOf<Blake2b256, Self> {
+        DigestOf::digest_byteslice(
+            &ByteBuilder::new()
+                .bytes(self.vrf_public_key.as_ref())
+                .bytes(self.kes_public_key.as_ref())
+                .finalize()
+                .as_byteslice(),
+        )
+    }
+}
+
+impl Readable for GenesisPraosLeader {
+    fn read<'a>(buf: &mut ReadBuf<'a>) -> Result<Self, ReadError> {
+        let vrf_public_key = deserialize_public_key(buf)?;
+        let kes_public_key = deserialize_public_key(buf)?;
+        Ok(GenesisPraosLeader {
+            vrf_public_key,
+            kes_public_key,
+        })
+    }
+}
+
 #[cfg(any(test, feature = "property-test-api"))]
 mod tests {
     use super::*;
+    use chain_crypto::{testing, Curve25519_2HashDH, PublicKey, SecretKey, SumEd25519_12};
+    use lazy_static::lazy_static;
     use quickcheck::{Arbitrary, Gen};
 
     impl Arbitrary for Hash {
@@ -324,6 +358,22 @@ mod tests {
                 EitherEd25519SecretKey::Normal(Arbitrary::arbitrary(g))
             } else {
                 EitherEd25519SecretKey::Extended(Arbitrary::arbitrary(g))
+            }
+        }
+    }
+    impl Arbitrary for GenesisPraosLeader {
+        fn arbitrary<G: Gen>(g: &mut G) -> Self {
+            lazy_static! {
+                static ref PK_KES: PublicKey<SumEd25519_12> =
+                    { testing::static_secret_key::<SumEd25519_12>().to_public() };
+            }
+
+            let tcg = testing::TestCryptoGen::arbitrary(g);
+            let mut rng = tcg.get_rng(0);
+            let vrf_sk: SecretKey<Curve25519_2HashDH> = SecretKey::generate(&mut rng);
+            GenesisPraosLeader {
+                vrf_public_key: vrf_sk.to_public(),
+                kes_public_key: PK_KES.clone(),
             }
         }
     }
