+++ chain-impl-mockchain/src/ledger/ledger.rs
@@ -12,8 +12,8 @@ use crate::config::{self, ConfigParam};
 use crate::date::{BlockDate, Epoch};
 use crate::fee::{FeeAlgorithm, LinearFee};
 use crate::fragment::{BlockContentHash, BlockContentSize, Contents, Fragment, FragmentId};
-use crate::leadership::genesis::ActiveSlotsCoeffError;
 use crate::rewards;
+use crate::setting::ActiveSlotsCoeffError;
 use crate::stake::{PercentStake, PoolError, PoolStakeInformation, PoolsState, StakeDistribution};
 use crate::transaction::*;
 use crate::treasury::Treasury;
