+++ chain-impl-mockchain/src/update.rs
@@ -1,8 +1,8 @@
 //use crate::certificate::{verify_certificate, HasPublicKeys, SignatureRaw};
 use crate::date::BlockDate;
 use crate::fragment::config::ConfigParams;
-use crate::leadership::{bft, genesis::ActiveSlotsCoeffError};
-use crate::setting::Settings;
+use crate::leadership::bft;
+use crate::setting::{ActiveSlotsCoeffError, Settings};
 use chain_core::mempack::{ReadBuf, ReadError, Readable};
 use chain_core::property;
 use chain_crypto::Verification;
